adding new upload lib to the scripts
